% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binHelper.R
\name{bin2gray}
\alias{bin2gray}
\title{A gray code converter function}
\usage{
bin2gray(x)
}
\arguments{
\item{x}{The binary number (base2) or a logical vector.}
}
\value{
The gray code as logical vector.
}
\description{
This function converts a binary number (base2) to a gray code
}
\seealso{
\link{gray2bin}
}

