% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binHelper.R
\name{binPrefix2Byte}
\alias{binPrefix2Byte}
\title{Binary prefix to Byte. (KiB,MiB,..)}
\usage{
binPrefix2Byte(n, prefix="KiB")
}
\arguments{
\item{n}{vector of numeric values}

\item{prefix}{binary prefix * Byte. Expeting a »string«}
}
\value{
The number of bytes fitting in n * binary prefix * Byte
}
\description{
Amount of Bytes needed to fit in n * KiB, MiB ..etc.
}
\details{
KiB <- KibiByte
MiB <- MebiByte
GiB <- GibiByte
TiB <- TebiByte
PiB <- PebiByte
EiB <- ExiByte
ZiB <- ZebiByte
YiB <- YobiByte
}
\examples{
binPrefix2Byte(c(0.5,1:10),"KiB")
as.binary(binPrefix2Byte(1,"KiB")*Byte())
}
\seealso{
\link{bytesNeeded} or \link{fillUpToByte} or \link{Byte}
}

