% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binHelper.R
\name{Byte}
\alias{Byte}
\title{A simple helper function to return the size of one Byte}
\usage{
Byte()
}
\value{
The size of one Byte (8)
}
\description{
Used to increase readabilaty
}
\seealso{
\link{bytesNeeded} or \link{addUpToByte} or \link{binPrefix2Byte}
}

