% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binOperators.R
\name{fillUpToByte}
\alias{fillUpToByte}
\title{Fill up to Byte (00000000..)}
\usage{
fillUpToByte(x, size=0, value=FALSE)
}
\arguments{
\item{x}{The binary number to fill up with zeros. (Any binary vector).}

\item{size}{in Byte. 0 = auto (smallest possible Byte).}

\item{value}{to fill up with FALSE(0) or fill up with TRUE(1).}
}
\value{
binary number. A binary vector with the desired size.
}
\description{
Fills up the binary number with zeros(0) or ones(1), to the size in Byte.
}
\details{
No floating point supported.
}
\examples{
fillUpToByte(as.binary(c(1,1), logic=TRUE), size=2)
fillUpToByte(as.binary(c(1,0,1), logic=TRUE), size=2, value=FALSE)
}
\seealso{
\link{fillUpToBit} or \link{bytesNeeded}, \link{negate}, \link{switchEndianess}.
}

