% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binHelper.R
\name{gray2bin}
\alias{gray2bin}
\title{A gray code to binary converter function}
\usage{
gray2bin(x, ...)
}
\arguments{
\item{x}{The gray code as logical vector.}

\item{...}{Additional parameter for binary()}
}
\value{
The binary number (base2).
}
\description{
This function converts a gray code to a binary number (base2)
}
\seealso{
\link{bin2gray}
}

