% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryS3.R
\name{is.binary}
\alias{is.binary}
\title{is Binary Vector}
\usage{
is.binary(x)
}
\arguments{
\item{x}{object to test.}
}
\value{
TRUE or FALSE.
}
\description{
test for object "binary".
}
\seealso{
\link{as.binary} and \link{binary}
}

