% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryS3.R
\name{loadAttributes}
\alias{loadAttributes}
\title{loadAttributes
 
Helper function load Attributes}
\usage{
loadAttributes(x, l)
}
\arguments{
\item{x}{x}

\item{l}{l}
}
\description{
loadAttributes
 
Helper function load Attributes
}

