\name{print.binary}
\alias{print.binary}
\title{Print method for binary number.}
\usage{
\method{print}{binary}(x, ...)
}
\arguments{
  \item{x}{any binary number.}

  \item{...}{further arguments.}
}
\value{
Output in ones and zeros (binary vector).
}
\description{
This method prints the binary number.
}
\seealso{
\link{summary.binary} provides some additional information.
}

