% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryS3.R
\name{saveAttributes}
\alias{saveAttributes}
\title{saveAttributes
 
Helper function save Attributes}
\usage{
saveAttributes(x)
}
\arguments{
\item{x}{x}
}
\description{
saveAttributes
 
Helper function save Attributes
}

