\name{as.binary}
\alias{as.binary}
\title{as Binary Vector}
\usage{
  as.binary(x)
}
\arguments{
  \item{x}{object to convert.}

  \item{signed}{TRUE or FALSE. Unsigned by default. (two's
  complement)}

  \item{littleEndian}{if TRUE. Big Endian if FALSE.}
}
\value{
  a binary vector.
}
\description{
  convert object to "binary".
}
\seealso{
  base::as.logical , base::is.logical, base::as.integer
  base::raw
}

