% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binAdd.R
\name{binAdd}
\alias{binAdd}
\title{Binary Addition (+)}
\usage{
binAdd(x, y)
}
\arguments{
\item{x}{summand 1 (binary vector)}

\item{y}{summand 2 (binary vector)}
}
\value{
The sum of x and y. Returns a binary vector.
}
\description{
Adds two binary numbers. (x + y)
}
\details{
Little-Endian and unsigned is not supported at the moment. No floating point supported.
if x or y is signed the return value will also be signed.
}
\examples{
five <- as.binary(5); ten <- as.binary(10);
as.numeric(binAdd(ten, five))
binAdd(as.binary(c(0,1), logic=TRUE), as.binary(c(1,0), logic=TRUE))
}
\seealso{
base::as.logical , base::is.logical, base::raw
}

