\name{binary}
\alias{binary}
\title{Binary Vector}
\usage{
  binary(n)
}
\arguments{
  \item{n}{length of vector.}
}
\value{
  a binary vector of length n
}
\description{
  Create objects of type "binary".
}
\details{
  No floating point supported.
}
\seealso{
  base::as.logical , base::is.logical, base::as.integer
  base::raw
}

