% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binOperators.R
\name{binSeq}
\alias{binSeq}
\title{Binary sequence}
\usage{
binSeq(x, ...)
}
\arguments{
\item{x}{a sequence.}

\item{...}{used for dec2bin().}
}
\value{
a sequence list of binary digits.
}
\description{
Binary sequence.
}
\examples{
binSeq(0:4)
}
\seealso{
\link{binary}
}

