% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binOperators.R
\name{rotate}
\alias{rotate}
\title{Rotate no carry ()}
\usage{
rotate(x, n)
}
\arguments{
\item{x}{The binary number to rotate. (binary or logical vector).}

\item{n}{The number of bits to rotate.}
}
\value{
rotates the vector from left to right. 
The value from MSB is used to fill up the vector at LSB. Returns a binary/logical vector.
}
\description{
A circular shift
}
\examples{
x <- as.binary(c(1,0,0,1,1,1,0,1), logic=TRUE); x
rotate(x,1)
rotate(x,2)
}
\seealso{
\link{shiftLeft} and \link{shiftRight}
}

