% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binOperators.R
\name{shiftRight}
\alias{shiftRight}
\title{Binary Right Shift (>>)}
\usage{
shiftRight(x, n)
}
\arguments{
\item{x}{The binary number to shift. (binary or logical vector).}

\item{n}{The number of bits to shift.}
}
\value{
Pushes 0's(FALSE) to the vector from left(MSB) to right(LSB).
Everything on right(LSB) side drops out. Returns a binary/logical vector
}
\description{
Logical right shift 1 >> n
}
\examples{
x <- as.binary(c(1,0,0,1,1,1,0,1), logic=TRUE); x
shiftRight(x,1)
shiftRight(x,2)
}
\seealso{
\link{shiftLeft} and \link{rotate}
}

