% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binOperators.R
\name{switchEndianess}
\alias{switchEndianess}
\title{Switch Endianess.}
\usage{
switchEndianess(x, stickyBits=FALSE)
}
\arguments{
\item{x}{binary number. Any binary number.}

\item{stickyBits}{Bits wont change if set TRUE. Only the attribute will be switched.}
}
\value{
switch little-endian to big-endian and vice versa.
}
\description{
Switch little-endian to big-endian and vice versa.
}
\examples{
x <- as.binary(c(1,1,0,0), logic=TRUE); print(x); summary(x);
y <- switchEndianess(x); print(y); summary(y);
y <- switchEndianess(x, stickyBits=TRUE); print(y); summary(y);
}
\seealso{
\link{negate} or \link{fillUpToByte}.
}

