% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binOperators.R
\name{switchEndianess}
\alias{switchEndianess}
\title{Switch Endianess.}
\usage{
switchEndianess(x)
}
\arguments{
\item{x}{binary number. Any binary vector.}
}
\value{
switch little-endian to big-endian and vice versa.
}
\description{
switch little-endian to big-endian and vice versa.
}
\examples{
x <- as.binary(c(1,1,0,0), logic=TRUE); print(x); summary(x);
y <- switchEndianess(x); print(y); summary(y);
neg_two <- as.binary(-2)
as.raw(neg_two)
as.raw(switchEndianess(neg_two))
}
\seealso{
\link{negate} or \link{fillUpToByte}.
}

